add working frontend

+++ backend/app/audio_analysis.py
@@ -126,20 +126,19 @@ def analyze_audio(file_path):
         band_energy = np.mean(S[mask])
         band_energies[band] = band_energy
 
-
     return {
-        "peak_db": round(peak_db, 2),
-        "rms_db": round(rms_db, 2),
-        "tempo": round(tempo, 2),
+        "peak_db": float(round(peak_db, 2)),
+        "rms_db": float(round(rms_db, 2)),
+        "tempo": float(round(tempo, 2)),
         "key": key,
-        "lufs": round(loudness, 2),
-        "dynamic_range": round(dynamic_range, 2),
-        "stereo_width_ratio": round(width_ratio, 3),
+        "lufs": float(round(loudness, 2)),
+        "dynamic_range": float(round(dynamic_range, 2)),
+        "stereo_width_ratio": float(round(width_ratio, 3)),
         "stereo_width": stereo_width_label,
-        "low_end_energy_ratio": round(normalized_low_end, 3),
+        "low_end_energy_ratio": float(round(normalized_low_end, 3)),
         "bass_profile": bass_profile,
         "band_energies": json.dumps({k: round(float(v), 2) for k, v in band_energies.items()}),
-        "issues": json.dumps(["issues"]) # json can take more than one issue
+        "issues": json.dumps(["issues"])
     }
